Added a shared constants file
